requirements.txt
re
xml
base64
riposte
pywinrm
requests
